Move custom prefs to separate package.
We have enough of those to justify keeping them separately.
