+++ mobile/src/main/java/org/openhab/habdroid/ui/preference/CustomInputTypePreference.kt
@@ -1,4 +1,4 @@
-package org.openhab.habdroid.ui.widget
+package org.openhab.habdroid.ui.preference
 
 import android.content.Context
 import android.util.AttributeSet
